Use IT suffix replace ITCase
